Adds a check for a nil 200 response body property.
If this is set, it does not mean that the response was actually necessarily empty. In most cases I've seen, Firefly III will respond in such a way (I suspect with redirects, but I have not investigated) that causes the client to get a 200 code response but the body is the HTML of the login page.

When I have seen this I have checked the Firefly III error logs and found a validation error of one sort or another.

In the future if I encounter this again I will investigate the redirect theory and see if I can prevent the client from following redirects to surface the true problem.